Add files via upload
when you step on the sensor, the capacitance about doubles. however, sometimes the starting value is different (starts at 5k pF and doubles to 10k, other times starts at 10k and doubles to 20k) why is this? does it matter? could write code so that instead of above a threshold value, it's when the capacitance changes significantly?